+++ src/main/java/musiccube/controllers/GenreController.java
@@ -1,15 +1,14 @@
 package musiccube.controllers;
 
-import musiccube.services.genre.GenreService;
 import musiccube.entities.Genre;
+import musiccube.entities.Song;
+import musiccube.services.genre.GenreService;
+import musiccube.services.song.SongService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.view.RedirectView;
-import springfox.documentation.annotations.ApiIgnore;
 
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
@@ -18,48 +17,66 @@ import java.util.Optional;
 
 @RestController
 @RequestMapping("/api")
-@CrossOrigin(origins = "http://localhost:4200")
+@CrossOrigin(origins = "${serverAddress}")
 public class GenreController {
 
     @Autowired
     private GenreService genreService;
 
-    @RequestMapping(value = "/genre{id}",
-            method = RequestMethod.GET,
-            produces = MediaType.APPLICATION_JSON_VALUE)
-    public Optional<Genre> getById(int id) {
-        return genreService.getById(id);
+    @Autowired
+    private SongService songService;
+
+    @GetMapping(
+            path = "/genre/{id}",
+            produces = MediaType.APPLICATION_JSON_VALUE
+    )
+    public ResponseEntity<Genre> getById(@PathVariable("id") int id) {
+        Optional<Genre> genre = genreService.getById(id);
+        return genre.isPresent() ?
+                ResponseEntity.ok(genre.get()) :
+                ResponseEntity.notFound().build();
     }
 
-    @RequestMapping(value = "/genres",
-            method = RequestMethod.GET,
-            produces = MediaType.APPLICATION_JSON_VALUE)
+    @GetMapping(
+            path = "/genre",
+            produces = MediaType.APPLICATION_JSON_VALUE
+    )
     public Iterable<Genre> getAll() {
         return genreService.getAll();
     }
 
     // --- Get all genres with paging ---
-    @RequestMapping(value = "genres/{page}",
-            method = RequestMethod.GET,
-            produces = MediaType.APPLICATION_JSON_VALUE)
-    public Iterable<Genre> getAllPaging(@PathVariable("page") Integer pageNr, @RequestParam("size") Optional<Integer> perPage) {
-        return genreService.getAllPaging(pageNr,perPage.orElse(10));
+    @GetMapping(
+            path = "genre/page/{page}",
+            produces = MediaType.APPLICATION_JSON_VALUE
+    )
+    public Iterable<Genre> getAllPaging(
+            @PathVariable("page") Integer pageNr,
+            @RequestParam(name = "size", required = false) Integer perPage
+    ) {
+        return perPage == null ?
+                genreService.getAllPaging(pageNr, Defaults.PAGESIZE) :
+                genreService.getAllPaging(pageNr, perPage);
     }
 
-    @RequestMapping(value = "/genres{name}",
-            method = RequestMethod.GET,
+    // --- Get by name ---
+    @GetMapping(
+            path = "/genre/name/{name}",
             produces = MediaType.APPLICATION_JSON_VALUE)
-    public Iterable<Genre> getByGenreName(String genreName) {
+    public Iterable<Genre> getByGenreName(@PathVariable("name") String genreName) {
         return genreService.getByGenreName(genreName);
     }
 
-    @RequestMapping(value = "/genre",method = RequestMethod.POST)
+    @PostMapping("/admin/genre")
     public ResponseEntity<Genre> create(@RequestBody @Valid @NotNull Genre genre) {
+        if (genreService.existsByName(genre.getGenreName())) {
+            return new ResponseEntity<>(HttpStatus.CONFLICT);
+        }
         genreService.save(genre);
         return ResponseEntity.ok().body(genre);
     }
 
-    @RequestMapping(value = "/genre",method = RequestMethod.PUT)
+    @PutMapping("/admin/genre")
     public ResponseEntity<Void> edit(@RequestBody @Valid @NotNull Genre genre) {
         Optional<Genre> genre1 = genreService.getById(genre.getId());
         if (Objects.nonNull(genre1)) {
@@ -68,15 +85,17 @@ public class GenreController {
         } else return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
 
-    @ApiIgnore
-    @RequestMapping(value = "/genres",method = RequestMethod.DELETE,produces = MediaType.APPLICATION_JSON_VALUE)
-    public Iterable<Genre> redirect(Model model) {
-        return genreService.getAll();
-    }
-
-    @RequestMapping(value = "/genre/{id}", method = RequestMethod.DELETE)
-    public RedirectView delete(@PathVariable Integer id) {
-        genreService.delete(id);
-        return new RedirectView("/api/genres",true);
+    @DeleteMapping("/admin/genre/{id}")
+    public ResponseEntity<Void> delete(@PathVariable Integer id) {
+        Genre genre = genreService.getById(id).orElse(null);
+        if(genre != null) {
+            Iterable<Song> allSongsWithGenre = songService.getByGenreName(genre.getGenreName());
+            for (Song song : allSongsWithGenre){
+                song.setGenre(null);
+                songService.save(song);
+            }
+            genreService.delete(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
 }
