+++ src/main/java/musiccube/controllers/InstrumentController.java
@@ -1,15 +1,16 @@
 package musiccube.controllers;
 
+import musiccube.entities.ArtistInstrument;
 import musiccube.entities.Instrument;
+import musiccube.entities.SongInstrument;
+import musiccube.services.artistinstrument.ArtistInstrumentService;
 import musiccube.services.instrument.InstrumentService;
+import musiccube.services.songinstrument.SongInstrumentService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.view.RedirectView;
-
 
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
@@ -18,44 +19,66 @@ import java.util.Optional;
 
 @RestController
 @RequestMapping("/api")
-@CrossOrigin(origins = "http://localhost:4200")
+@CrossOrigin(origins = "${serverAddress}")
 public class InstrumentController {
     @Autowired
     private InstrumentService instrumentService;
 
-    @RequestMapping(value = "/instrument{id}",
-            method = RequestMethod.GET,
-            produces = MediaType.APPLICATION_JSON_VALUE)
-    public Optional<Instrument> getById(int id) {
-        return instrumentService.getById(id);
+    @Autowired
+    private SongInstrumentService songInstrumentService;
+
+    @Autowired
+    private ArtistInstrumentService artistInstrumentService;
+
+    @GetMapping(
+            path = "/instrument/{id}",
+            produces = MediaType.APPLICATION_JSON_VALUE
+    )
+    public ResponseEntity<Instrument> getById(@PathVariable("id") int id) {
+        Optional<Instrument> instrument = instrumentService.getById(id);
+        return instrument.isPresent() ?
+                ResponseEntity.ok(instrument.get()) :
+                ResponseEntity.notFound().build();
     }
 
-    @RequestMapping(value = "/instruments",
-            method = RequestMethod.GET,
-            produces = MediaType.APPLICATION_JSON_VALUE)
+    @GetMapping(
+            path = "/instrument",
+            produces = MediaType.APPLICATION_JSON_VALUE
+    )
     public Iterable<Instrument> getAll() {
         return instrumentService.getAll();
     }
     
     // --- Get all instruments with paging ---
-    @RequestMapping(value = "instruments/{page}",
-            method = RequestMethod.GET,
-            produces = MediaType.APPLICATION_JSON_VALUE)
-    public Iterable<Instrument> getAllPaging(@PathVariable("page") Integer pageNr, @RequestParam("size") Optional<Integer> perPage) {
-        return instrumentService.getAllPaging(pageNr,perPage.orElse(10));
+    @GetMapping(
+            path = "instrument/page/{page}",
+            produces = MediaType.APPLICATION_JSON_VALUE
+    )
+    public Iterable<Instrument> getAllPaging(
+            @PathVariable("page") Integer pageNr,
+            @RequestParam(name = "size", required = false) Integer perPage
+    ) {
+        return perPage == null ?
+                instrumentService.getAllPaging(pageNr, Defaults.PAGESIZE) :
+                instrumentService.getAllPaging(pageNr, perPage);
     }
 
     // --- Get by instrument name ---
-    @RequestMapping(value = "/instruments{name}",method = RequestMethod.GET,produces = MediaType.APPLICATION_JSON_VALUE)
-    public Iterable<Instrument> getByInstrumentName(String instrumentName) { return instrumentService.getByInstrumentName(instrumentName); }
+    @GetMapping(
+            path = "/instrument/name/{name}",
+            produces = MediaType.APPLICATION_JSON_VALUE
+    )
+    public Iterable<Instrument> getByInstrumentName(@PathVariable("name") String instrumentName) {
+        return instrumentService.getByInstrumentName(instrumentName);
+    }
 
-    @RequestMapping(value = "/instrument",method = RequestMethod.POST)
+    @PostMapping("/admin/instrument")
     public ResponseEntity<Instrument> create(@RequestBody @Valid @NotNull Instrument instrument) {
         instrumentService.save(instrument);
         return ResponseEntity.ok().body(instrument);
     }
 
-    @RequestMapping(value = "/instrument",method = RequestMethod.PUT)
+    @PutMapping("/admin/instrument")
     public ResponseEntity<Void> edit(@RequestBody @Valid @NotNull Instrument instrument) {
         Optional<Instrument> artist1 = instrumentService.getById(instrument.getId());
         if (Objects.nonNull(artist1)) {
@@ -64,14 +87,20 @@ public class InstrumentController {
         } else return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
 
-    @RequestMapping(value = "/instrument",method = RequestMethod.DELETE,produces = MediaType.APPLICATION_JSON_VALUE)
-    public Iterable<Instrument> redirect(Model model) {
-        return instrumentService.getAll();
-    }
-
-    @RequestMapping(value = "/instrument/{id}", method = RequestMethod.DELETE)
-    public RedirectView delete(@PathVariable Integer id) {
-        instrumentService.delete(id);
-        return new RedirectView("/api/instrument",true);
+    @DeleteMapping("/admin/instrument/{id}")
+    public ResponseEntity<Void> delete(@PathVariable Integer id) {
+        Instrument instrument = instrumentService.getById(id).orElse(null);
+        if(instrument != null) {
+            Iterable<ArtistInstrument> allArtistsWithInstrument = artistInstrumentService.getByInstrumentId(instrument.getId());
+            Iterable<SongInstrument> allSongsWithInstrument = songInstrumentService.getByInstrumentId(instrument.getId());
+            for (ArtistInstrument artistInstrument : allArtistsWithInstrument){
+                artistInstrumentService.delete(artistInstrument.getId());
+            }
+            for (SongInstrument songInstrument : allSongsWithInstrument) {
+                songInstrumentService.delete(songInstrument.getId());
+            }
+            instrumentService.delete(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
 }
